requirements.txt
absl-py>=0.6.0
tensorflow
numpy>=1.15.2
scipy>=1.0.0
pandas>=0.24.1
scikit-image
tqdm